Expanded and genericized Ant build script. Completely backwards-compatible. If something you've been doing with this script stops working, please tell me.
However:
1) If you have CIP/lib/build.jar, it will run the code generators for
you (it finds the .def files automatically)
2) If we ever have files needing rmic, there are targets included (you
must specify the file pattern up front. **/*Impl.class works for me in
the core and server modules
3) New targets for creating a zip of the javadocs, a zip of the
runtime required files, and a zip of the development sources (a la
hourly builds)
4) Targets to handle compiling any examples directory
5) All targets defined such that module-specific settings are all set
way up front

The result of all this is that this Ant script can be easily used to
compile any Cougaar module. It does not compute dependencies or try to
compile dependant modules.