Try to really implement experiment copy: Copy the configuration as well. This is still incomplete - I cant get the Agents re-assigned to the Nodes.
Fix a couple bugs on experiments with one trial - must set
variationScheme and initialize the trial properties

Add in methods to check the experiment:
1) Does it have a configuration (hasConfiguration), ie does it have at
least one host with at least one node with at least one agent.
---- THIS METHOD IS BUGGY
2) Does it have any unbound properties (for use in deciding if it can
be run as-is
3) Create a default configuration: The local host with one node and
all of the agents.
--- This is useful to permit a user to run a society without having to
do anything in the experiment builder

Modify the meaning of isRunnable
-- It not checks to see if the Experiment has any
unBoundProperties. It should also check to see if it has a
configuration.
The result of this is that the Run menu item / console are not lit up
on an Experiment that has any unbound properties (those without a
value or a set of experiment values)
When it works, it should also check that the experiment has a
configuration

Another comment on something to fix: When an experiment is stopped,
tell the Societies that they are not running.