As I did for societies, make copy on Impacts use the stuff in ComponentProperties. This also means at the top of the file are the definitions of the classes of Impacts that are handled.
EXCEPT: I left in a special case copy of ABCImpact (the only one we
have) so that we can still copy it while Brian updates it to be a
ComponentProperties.

Note that as we transition to use this copy method, everything must
store all of its items in Properties, or else override the copy
method.