* Organizer.java: Add InsertionComponentRecipe to the list of built-in recipes.
* Organizer.java:
Added code to save recipes from the workspace to the database.
Added duplicate recipe name check when reading recipes from the
database suggesting that the workspace recipe be deleted or
renamed.

* Experiment.java:
Pass pdb to modifyComponentData in case it's needed.

* PopulateDb.java:
Add methods to give access to database for use by
ComponentInsertionRecipe.

* ConfigurableComponent.java:
Implement two argument variant of modifyComponentData by calling
the single argument variant.

* ComponentProperties.java:
Add variant of modifyComponentData accepting a PopulateDb to pass
in context to recipes. The default implementation just calls the
single argument variant so existing classes are not changed.

* PopulateDb.java (Repository):
Now extends PDbBase.

* org/cougaar/tools/csmart/recipe/ComponentInsertionRecipe.java:
New recipe component that inserts components based on the result
of doing queries.