* ExperimentTree.java: Fix is droppable test to allow dropping if _either_ the component is editable or if it has no unbound properties. Components without unbound properties can be safely shared.
* CMT.java:
Fix deleteExperiment. Code assumed that trial ids had a particular
form. That form did not match trial ids assigned when clonging an
experiment prior to adding host-node-agent mappings and recipes.
Fix is to query the db for the trial id of the experiment.

* PopulateDb.java:
Compute the "assemblyMatch" predicate so queries can readily test
if an assembly_id belongs to the current trial or not.

* ConfigurableComponent.java:
Add hasUnboundProperties() method so we can easily test if it's ok
to use share a component. Components with unbound properties can't
generally be shared because binding the properties in one context
would invalidate it in other contexts.

* ComponentProperties.java:
Add hasUnboundProperties() method to sense whether a Component has
any unbound properties.

* ComponentInsertionRecipe.java:
Remove debug println