Put LIBPATHS on org.cougaar.class.path as well as the system classpath. In most cases, this is entirely un-necessary.
However, if you are using stuff on COUGAAR_DEV_PATH this is otherwise
only loaded via the system classloader, and classes loaded from there
similarly are also only loaded via the system classloader. So, for
example, if you are developing in CSMART (so use _DEV_PATH and have no
jar file for the bootclassloader to find), then DBUtils.java will be
loaded via the system classloader. Therefore, your database driver
will be loaded via the system classloader. And if the appropriate jar
file is not on your system classpath, you'll get a
ClassNotFoundException