* Organizer.java: Add experiment name argument to deleteExperiment. Restore the properties of Host and Node components from the database. Change the getHosts query to only retrieve hosts that are part of the experiment instead of all the hosts in the database.
* Experiment.java:
Add the properties of nodes and hosts (Description, location, etc)
as parameters of the corresponding ComponentData so they will be
saved to the database.

* ExperimentDB.java:
Add experiment name argument to deleteExperiment call.

* CMT.java:
Add experiment name to deleteExperiment so that society components
created for the experiment can be deleted.

* PDbBase.java:
Make query logging option controlled by
csmart.PopulateDb.log.enable property. If true, all queries
executes are logged to a file named PopulateDb<datetime>.log
preceded by the time to execute the query.