Change from the branch:
Recipe: Added more required Property Groups, fixed a few bugs. Made
asset class a property, defaulted to MilitaryOrginazition. Added Org
Plugins.

PopulateDb.java Fixed Bugs related to populating an agent and its property groups, Fixed bug in populate agent
ConfigurableComponent.java Removed some printout chatter
PropertyEditorPanel.java Removed some printout chatter
Organizer.java: Print out the query when we get an SQLException, for debugging
 If we get no assemblies, produce an in string that is not an SQL
 Exception, and print a warning message


* HostConfigurationBuilder.java:
Fix bug 853
Use the default command line properties as the exclusive
definition of the nameserver property. Discard redundant (and
therefore incorrect) variables purporting to represent the same
value. If the setting of the nameserver property is absent or
specifies a host having no nodes with one or more agents, then
reset the nameserver property to the first host have a node with
at least one agent. As long as the nameserver property specifies a
host that the experiment controller will actually create a node
on, it will not be changed. If, however, the configuration is
changed such that the host specified by the nameserver property
will not be run for the experiment, the nameserver property is
changed to be a host that will be run. This insures that there
will be a nameserver, but allows any valid host to be used.

* CSMARTConsole.java:
Fix bug 853
Remove dead code involving nameserver settings.