Merge changes from 862 branch. See branch comments. Changes include: CSMART Console: Fixed bug 824 -- make "exception" be the default notification string. HostConfigurationBuilder: Ensure that node names and agent names are unique as they occupy the same name space.  Fixes bug 810. ui/console: Fixed bugs 771, 891, 893: Can provide local and global command line arguments from a file (use the "Command Line Arguments" and "Global Command Line Arguments" menu items on the node in ExperimentBuilder -- there's a new button in the pop-up dialog -- "Read From File"). Prevent user editing arguments that shouldn't be edited (such as node name).
* ReadOnlyProperties.java:
Fix bug 893
Extends java.util.Properties by adding a Set of keys that should
be immutable (or mutable only in a controlled way). The put,
putAll, remove, and clear methods are overridden to disallow
modifications or removal of the read-only keys. Additional methods
allow these values to be changed when needed.

* Experiment.java:
Fix bug 893
Change defaultNodeArguments to be ReadOnlyProperties. Make the
"...experiment.id" property be read-only and set its value
whenever the trialId is changed.
Simplify the copying of properties when copying the node.

* ExperimentNode.java:
Fix bug 893
Change "arguments" to be ReadOnlyProperties. Make the
"...node.name" property be readonly and set its value only when
renaming the node.
Simplify the copying of properties when copying the node.

* NodeComponent.java:
Fix bug 893
Remove the setArguments method. It is unnecessary and problematic
since it implies that the arguments of a node can be alter in a
will-nilly fashion. The only permitted actions are the editing of
the Properties returned by getArguments().

Experiment: Changed default node argument from:
org.cougaar.core.cluster.persistence.enabled
to
org.cougaar.core.cluster.persistence.enable (no final d)

If user exits Experiment Builder and hasn't modified the experiment,
then don't query about saving the experiment.
If user selects Save As menu item, then just save the experiment,
regardless of whether or not it's been modified.
These are additional fixes on bug 883.

CSMARTConsole: Fixed bug 851 -- if user doesn't specify results (metrics) directory,
then don't run the experiment (in ExperimentController).

CMT.java: Fix bug 885
Remove code that tries to remove the society component when
deleting an experiment. This violates foreign key constraints
against the component_hierarchy and component_arg tables. Let the
assembly cleanup code take care of this.

ui/viewer:
Allow re-use of experiment name when configuring an experiment (bug #883).
Experiment: Removed isCloned/setCloned methods; no longer used.
ExperimentBuilder: Added Save As menu item and method; call SaveAs method when user exits
ExperimentBuilder (by closing window or invoking File-Close menu item).
CMTDialog: Allow user to save experiment under original name.
CSMART: Pass through "allowExistingName" flag to getUniqueExperimentName.
Organizer: Added flag to getUniqueExperimentName to allow reuse of existing name.
Removed use of Experiment isCloned/setCloned.

ui/component:
* PDbBase.java:
* PopulateDb.java:
Fix bug 883.
Clone the experiment being saved if and only if its name is
different from the name in the database for the experiment
identified by the experiment id. Ignore arg saying whether to
createNew or not. If the name differs, it must be cloned. If the
name is the same it cannot be cloned.
Set experiment and trial descriptions rationally. Select
experiment and trial ids more reasonbably.

* Organizer.java:
Fix bug 882
When reading society and node properties replace all default
values with values from the database if there are _any_ values in
the database.

* HostConfigurationBuilder.java:
* Experiment.java:
More Fix bug 853
Fix problem in retaining nameserver host property. This
computation was done far too often including cases where the
complete list of hosts nodes and agents was not available.
Move the code to do this computation from builder to Experiment.
Do the computation before handing out the experiments list of
Nodes. This insures that any code that looks at the nodes'
properties will see an up=to=date computation of this property.


* Organizer.java:
Fix bug 878
Allow deletion if the recipe differs as well as if it exists.

* Experiment.java:
Fix bug 876
Save agent properties. This allows recipe added agents to be
marked with a property so they can be identified when read from
the database. This allows old recipe agents to be removed.

* PopulateDb.java:
Fix bug 876
Provide a way to repopulate the CMT part of an experiment. The CMT
part is normal read-only, but if a recipe deletes some part of the
CMT assembly, the only way to represent that is to replace the old
CMT with a new omitting or changing as necessary.

* PDbBase.java:
Fix bug 876
Save only local properties for a recipe.

* GenericComponentData.java:
Fix bug 876
setChildren must clear old children before setting new ones
otherwise they just keep building up.

* ConfigurableComponent.java:
* ComponentProperties.java:
Fix bug 876
Provide mechanism to indicate removed (renamed) agents.

* PopulateDb.java:
Fix bug 875
Improve detection of inserted/replaced arguments to not falsely
indicate removal when none has happened.

* CMTAgent.java:
Fix bug 876
Allow construction without assemblyId for the case where the agent
has not yet been put in the database.

Fixed Bug: #873
Agent Insertion Recipe now provides the ability to
toggle the inclusion of OrgAssets and ItemIdentificationPG's.
If an OrgAsset is Included, a ClusterPG is created as well.
ItemIdentificationPG's are required for all agents except
Prototypes.

Fixed bug #841 -- allow forced editing of a recipe that is included
in an experiment that has been run.