Fixes from branch:
Multiple changes:
1) Fix addition of recipe components to ensure order is as listed in
ExperimentBuilder, components arent duplicated, and binders come
before other items at same level.
Note you cannot have 2 plugins with the same name in an agent, binders
always come before _every_ plugin (although the infrastructure would
let you do otherwise). Also note that earlier recipes _will_ apply to
later recipes: If an earlier recipe adds a plugin to all agents, and a
later recipe adds an agent, the new agent _will_ have the new plugin.
Also note that if you edit an Agent Recipe that is in an experiment,
when you load the experiment back from the database, the recipe
appears to have been removed from the experiment.

2) New debugging tool: All ini files can be dumped to the results
   directory. This includes showing the -D arguments for the society,
   and the per-node arguments. Note that the prototype-ini files are
   not complete for society agents.


Organizer: When creating Experiment from DB, add in agents from
society before those from recipes, and only add in the recipe agents
if they're not already in the list

GenericComponentData/LeafComponentData: add a .equals method: 2
ComponentData are equals if they're names are equals.

ComponentData/GenericComponentData: new convenience method to give the
index of a child (-1 if not present)
-- new convenience method used in adding children: always replace any
existing child that is .equals (same name). Binders go before all
agents/plugins. Otherwise add new items after others of the same type.

HostConfigurationBuilder: The ConfigurableComponentComparator should
compare the short name, not the full name when comparing nodes and
agents with DB societies: It is used for Nodes (names are only short),
and Agents (whose names can't be duplicated anyhow). Note the solution
in here hard-codes the society prefix of Combo for DB societies - bug 914

CMTAgent: When grabbing the set of assemblies for constructing
ComponentData, only use assemblies that start with CMT - avoids
constructing stuff that Recipes should be adding. Also add some
comments. Also, add a .equals method: Agents are the same if their
short name is the same. Note that this is not strictly true (built in
societies' agents need to compare the longer name)

AgentInsertionRecipe: Most references to InsertAgentComponent need not
do so, change them to a generic AgentComponent.
- When creating the the AgentComponentData, create it fresh and
replace the existing one. This forces it to be all the new stuff. It
has the side effect that recipes listed before this one will never
effect this agent
- Comment out some debug printlns
- Change .equals method to allow comparison with any AgentComponent,
and
- to only compare on the short name. This won't work quite right with
built-in societies, which make use of the longer agent name

ABCImpact & BasicMetric & ComponentInsertionRecipe & SpecificInsertionRecipe: use the new
addChildDefaultLoc method for adding children in the correct location.

ComponentInsertionRecipe: If the type I get from the query is binder,
treat it as an agent binder (without this, type of component is wrong)

PopulateDb.java:
When trying to get the componentAlibID for a component, handle the
case where it has no ancestor: use the name as a default, but dont set
the ID on the ComponentData - let someone else handle that (causes
NPEs when adding Agents)

ExperimentBuilder/Experiment: new File menu item - a debug option to
print out all the .ini files locally. Note that the OrgAsset files are
not complete. This tries to work for both DB and non-DB
experiments. For both, it only writes out the configs for the first
trial (untested for non-DB experiments that really vary parameters).
- helps address bug 915, 916

ExpConfigWriterNew: Changes to support writing out debug ini files
- New constructor that takes full ComponentData
- Build the non-DB component data strictly forward (not reverse for
calling modifyComponentData as before)
- Catch NPE when the Nodes don't have a ConfigurationFileName property
(which they don't for DB experiments)
- Handle the possibility that below the top-level society node, there
may be host nodes. Write out a file for each host, listing its Nodes +
the node-specific arguments. Also write out a society file, listing the global -D arguments
- When writing individual lines like plugin = ..., use the ClassName
in many circumstances
- Print a stacktrace on exceptions
- Translate node binder and agent binder into the appropriate
insertion point
- Dont write the prototype ini file at all if we have no relevant data
- Dont write the UniqueID slot in the prototype ini if its null
These are bugs 915, 916

ABCImpact: Change Node Binder insertion point to
Node.AgentManager.Binder