Fix for bug 921. Core allows 2 plugins (or binders, etc) of same class in same agent. To allow this in CSMART, construct alib_ids using the component _name_, not the Class. Users must now specify unique names for plugins/binders they add with recipes, in order to add, for example, the Servlet plugin twice to the same agent.
Note that if the arguments to the plugin are identical, Cougaar may
have trouble, but CSMART will currently allow this.

It is still not possible to modify or delete a binder/plugin/agent
using a recipe.

If you try to add a plugin to an agent twice, using the same component
name, you will get an IllegalArgumentException.

Also:
ParameterInsertionRecipe correctly compares Alib_id's with Alib_ids,
or classes with classes (and not names with alib ids, which makes no
sense).

Additionally, added a comment of a possible improvement to
GenericComponentData