Fairly substantial modifications:
The /data/common/servlets.txt file now expects the class for the
loader of the servlet, as well as for the servlet itself.

As a result, we can now use this file to handle servlets that are
their own loader, as with the /load servlet.

Added use of the logger
Allow comment lines and blank lines in servlets.txt

Note that although this is called ServletGroupInsertion, it is really
PluginGroupInsertion -- to make it really be this would mostly just
require cosmetic changes. I bet it could be used for that as is.