Very numerous changes, many only partially tested, including: Consolidate constructors: there is now only one for use with experiments. It takes a societyID as an argument. This society will be saved in the runtime area/ config area if not already there (and other older CMT/CSA assemblies will be removed)
New helper method to get the type of an assembly_id

New helper method to get the current description of an assembly

Added lots of debug output

New helper methods to delete all CMT / CSA assemblies from config /
runtime

New _static_ method to delete a given assembly if it is not currently
used anywhere. Uses the simple constructor. To support this, made the
local Logger reference public

Helper method to see if the given assembly is in the Config table for
this Trial. Ditto for the runtime table

Helper method to add an assembly to the runtime table

Filled in populateCSA. It now will also, if we are saving an experiment 1) creates a CSA assembly, 2) deletes
any old runtime assemblies that might exist, 3) adds the new CSA
assembly to the runtime table, and then 4) calls populate
-- the result is that this method can be called instead of the old
repopulateCMT

Fix up the check on new Agents:
Make sure the new item is an Agent before marking it as new.
Also, a big comment noting the limitations of this check.

Change the way attribute_orders are set for PG Values: Make them start
at 0, as is done in the database. This way we avoid lots of
duplicates, etc.

Also, check that the value is not already there (in one of the runtime
assemblies) before inserting a new PG value

Fix up the RelationshipData stuff: the Role in the RelationshipData is
often empty, in which case it appears that Type is the slot you
want. FIXME: This may not be right. It may be we just want the literal
Subordinate

Fix up the Relationship handling of star/end date: These are not
question marks. Instead, get the Long values from the RelationshipData
object. But use null for the end time if we have no value, not 0.