Tons more debug output Re-work cleanTrial, trying to fix my little bug....
Fix in saving PG multi-vals to DB. The CMT societies dont use
argument_order when getting the pg multi-vals from the DB, but we
do. This means we'll insert duplicates when we dont need to.
So Ive added a check: If there is more than one value for the PG slot,
and all the values are numbered with 0, and my value is already there,
then dont add it again.