Log error if have null assembly in various populate methods
When saving an experiment in a CSA assembly (ie there was a
modification) do NOT completely get rid of the CSMI assembly (since we
may later still try to save there anyhow). Just make sure it is
cleaned out before we continue.

Note the significance: I have not checked fixAssemblies to see if it
agrees, but the idea is that in these removal/modification situations
you still might end up with both a CSA and a CSMI assembly at runtime.

I ran into this when, during the save of the real society, 2 Arguments
were not equal (which should have been). Why were they not equal?
Don't know. But this registered as a modification, and therefore I
ended up in populateCSA, and later called populateCSMI, and discovered
my bug.