turn down some logging.
If a node has no children, and is not assigned to a Host, do not write
it out. This means that if you are dumping INIs for a configuration
where you have not assigned Nodes to Hosts, any completely empty Nodes
will not be written out. I _think_ this is OK.

However, it is still currently true that if you think you have nicely
completely configured your experiment, but have some generic
un-assigned Nodes listed, and have some recipe that adds things to
Nodes (like the Mobility recipe), then these generic Nodes will not be
empty, and they will be written out - which looks confusing.

We could avoid writing these out if they have no Agents maybe?