Put an Attach menu item on the AppServer drop-down menu, same behavior as the button
Put in some logic to try to enable / disable various appServer actions
based on whether there are AppServers or processes

Put up a confirmation dialog on the Kill All Nodes command

If have an experiment, not run, attach to a single node, kill it, then
select re-start -- this should not be the same as hitting Run - cause
that would run the Experiment.

If hit Attach, and there are no New Nodes to Attach To, say so, don't
just do nothing

Try to enable the Run button when start with an experiment and have
attached to some Nodes

Run / stop / etc buttons not quite enabled at right times yet - still
working on it.