Use core constants for insertion points, not string literals
Warning: Note that on save, removal of a PG or PG slot is not noticed:
so the old PG / slot will be present at runtim!

Hack for 10.0 upgrade: If we have a MessageAddress, and have a
ClusterIdentifier in the DB, then we just did an upgrade - resave as a
CSA, so that at runtime, we have no ClusterIdentifier.
Warning: this _appears_ to put the CSA in the config as well as
runtime -- is this right?