First pass save optimization: ComponentData track when they have been modified. If a ComponentData is modified, it's parents are marked modified (recursing up). (AgentAssetData are treated like ComponentData for this purpose.)
When you save a CSMI assembly, you are looking for just the
changes. So by looking at this modified flag, we know if any recipe
touched this part of the CData tree, without having to compare with
the DB contents.
Note this _only_ works for the CSMI assemblies: CSHNA and CMT/CSA, as
well as RCP assemblies are generated by a single component, so
anything there is new and needs to be saved.

Note also that to make this work, after any society does
addComponentData, we must clear the modified flag
(societyCData.resetModified() which recurses down), so that any recipe
additions in addComponentData are seen as changes. This is done in
SocietyBase.
In addition, each of the parent populate methods in PopulateDb.java
resets the modified flags after saving an assembly
(CSMI/CSHNA/RCP/CSA).

Result:
When saving the CSMI assembly, it does not recurse into most
AgentAssetData or Plugin/Binder CDatas. This should reduce the DB
interactions nicely.
It also means that if we undid the change to save all recipe changes
at once, we might find the performance difference was small.