Allow GLS Servlet to be either on NCA or on OSD.GOV
Only look for GLS Servlet if you have an experiment

If you can't find GLS Servlet, avoid NPE

If you find the agent with the GLS Servlet, use the Node Model /
experiment to try to intuit the correct protocol, host, and port --
and preserve the agent on which you found the servlet to use when
constructing the GLS Init UI

Also use this smarts when user manually asked for the GLS Init UI from
the menu item.

Warning:
1) We do not actually search for the GLS Init servlet, but assume that
if we find an NCA or OSD.GOV agent, then the servlet is there
2) When you stop your experiment, you'll get a warning dialog from the
GLS Client saying you have a bad connection. If we kept a pointer to
the client someplace, we could theoretically call glsClient.stop()
first